Don't minify Martor sources
Currently, the project ships with:

* a minified Martor JS file, which is not used; and
* minified as well as unminified Martor CSS files, where only the minified ones
  are used

This makes development awkward, since one has to remember to re-minify the CSS
files before testing changes, and is error-prone. Users concerned about the few
extra bytes being downloaded are likely using something like django-compressor
to begin with, which would take care of the minification for them.

This PR removes the unused minified JS file, and switches the widget to use the
unminified CSS sources.